backend/requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
python-dotenv==1.0.1
sqlparse==0.5.1
gunicorn==20.1.0
